requirements.txt
# Core data science and analysis
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0

# HTTP requests and web scraping
requests>=2.25.0
beautifulsoup4>=4.12.3

# Web automation and scraping
selenium>=4.23.0
webdriver-manager>=4.0.1

# Scheduling
schedule>=1.2.0

# Database
# sqlite3 is built-in with Python

# Optimization and machine learning
scikit-learn>=1.0.0

# Statistical analysis
statsmodels>=0.13.0

# Date/time utilities
python-dateutil>=2.8.0
pytz

# Web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Real-time data and async
websockets>=11.0.0
asyncio-mqtt>=0.11.0

# Environment variables
python-dotenv>=0.19.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Type hints
typing-extensions>=4.0.0